+++ libbpf-rs/CHANGELOG.md
@@ -10,6 +10,8 @@ Unreleased
   with custom file flags.
 - Changed name members of `query` info types to be `OsString` instead of
   `CString`
+- Made `query::KprobeMultiLinkInfo` and `query::UprobeMultiLinkInfo`
+  non-exhaustive
 
 
 0.26.2